Services/django/requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
defusedxml==0.7.1
Django==5.1.6
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
idna==3.10
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
django-storages
boto3
python-docx
openpyxl
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.5
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
whitenoise==6.9.0
